fix A LOT of compilation warnings

+++ ext/GraphCutOptimization/src/graph.cpp
@@ -16,8 +16,8 @@ template <typename captype, typename tcaptype, typename flowtype>
 	if (node_num_max < 16) node_num_max = 16;
 	if (edge_num_max < 16) edge_num_max = 16;
 
-	nodes = (node*) malloc(node_num_max*sizeof(node));
-	arcs = (arc*) malloc(2*edge_num_max*sizeof(arc));
+	nodes = (node*) malloc( (size_t) node_num_max * sizeof(node));
+	arcs = (arc*) malloc( 2 * (size_t) edge_num_max * sizeof(arc));
 	if (!nodes || !arcs) { if (error_function) (*error_function)("Not enough memory!"); exit(1); }
 
 	node_last = nodes;
@@ -66,7 +66,7 @@ template <typename captype, typename tcaptype, typename flowtype>
 
 	node_num_max += node_num_max / 2;
 	if (node_num_max < node_num + num) node_num_max = node_num + num;
-	nodes = (node*) realloc(nodes_old, node_num_max*sizeof(node));
+	nodes = (node*) realloc(nodes_old, (size_t) node_num_max * sizeof(node));
 	if (!nodes) { if (error_function) (*error_function)("Not enough memory!"); exit(1); }
 
 	node_last = nodes + node_num;
@@ -90,7 +90,7 @@ template <typename captype, typename tcaptype, typename flowtype>
 	arc* arcs_old = arcs;
 
 	arc_num_max += arc_num_max / 2; if (arc_num_max & 1) arc_num_max ++;
-	arcs = (arc*) realloc(arcs_old, arc_num_max*sizeof(arc));
+	arcs = (arc*) realloc(arcs_old, (size_t) arc_num_max * sizeof(arc));
 	if (!arcs) { if (error_function) (*error_function)("Not enough memory!"); exit(1); }
 
 	arc_last = arcs + arc_num;
