+++ ext/GraphCutOptimization/src/graph.h
@@ -268,9 +268,11 @@ private:
 								//   (or to itself if it is the last node in the list)
 		int			TS;			// timestamp showing when DIST was computed
 		int			DIST;		// distance to the terminal
-		int			is_sink : 1;	// flag showing whether the node is in the source or in the sink tree (if parent!=NULL)
-		int			is_marked : 1;	// set by mark_node()
-		int			is_in_changed_list : 1; // set by maxflow if 
+		// C bit-fields
+		// https://en.cppreference.com/w/c/language/bit_field
+		unsigned int			is_sink : 1;	// flag showing whether the node is in the source or in the sink tree (if parent!=NULL)
+		unsigned int			is_marked : 1;	// set by mark_node()
+		unsigned int			is_in_changed_list : 1; // set by maxflow if 
 
 		tcaptype	tr_cap;		// if tr_cap > 0 then tr_cap is residual capacity of the arc SOURCE->node
 								// otherwise         -tr_cap is residual capacity of the arc node->SINK 
@@ -375,7 +377,7 @@ template <typename captype, typename tcaptype, typename flowtype>
 	}
 	else
 	{
-		memset(node_last, 0, num*sizeof(node));
+		memset(node_last, 0, (size_t) num * sizeof(node));
 
 		node_id i = node_num;
 		node_num += num;
